% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc2AdjMat.R
\name{AdjMat2Graph}
\alias{AdjMat2Graph}
\title{AdjMat2Graph}
\usage{
AdjMat2Graph(adjMatrix)
}
\arguments{
\item{adjMatrix}{an adjacency matrix.}
}
\value{
a graph object of class \code{igraph}.
}
\description{
Create an undirected graph from an adjacency matrix of mvc
}
