% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarmeta.R
\name{coef.metavarmeta}
\alias{coef.metavarmeta}
\title{Estimated Parameter Method for an Object of Class
\code{metavarmeta}}
\usage{
\method{coef}{metavarmeta}(object, ...)
}
\arguments{
\item{object}{an object of class \code{metavarmeta}.}

\item{...}{further arguments.}
}
\value{
Returns a vector of estimated parameters.
}
\description{
Estimated Parameter Method for an Object of Class
\code{metavarmeta}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
