% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarmeta.R
\name{confint.metavarmeta}
\alias{confint.metavarmeta}
\title{Confidence Intervals for an Object of Class
\code{metavarmeta}}
\usage{
\method{confint}{metavarmeta}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{metavarmeta}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Confidence Intervals for an Object of Class
\code{metavarmeta}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
