% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan2mvc.R
\name{dbscan2mvc}
\alias{dbscan2mvc}
\title{dbscan2mvc}
\usage{
dbscan2mvc(dbscanRes, cov, mvcMinVar = 1000, mvcMinCov = 5)
}
\arguments{
\item{dbscanRes}{\code{dbscan} result.}

\item{cov}{metavaraint loci depth of coverage dataframe.}

\item{mvcMinVar}{minimum number of metavariants for mvc, set to 1,000 by default.}

\item{mvcMinCov}{minimum vamue for the median depth of coverage}
}
\value{
a list of object of class \code{mvc}
}
\description{
Convert a dbscan ouput to metavariant clusters
}
