% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-fit-ct-var.R
\name{FitCTVAR}
\alias{FitCTVAR}
\title{Fit First Order Continuous-Time Vector Autoregressive Model by ID}
\usage{
FitCTVAR(
  data,
  observed,
  id,
  time,
  phi_start = NULL,
  phi_lbound = NULL,
  phi_ubound = NULL,
  sigma_start = NULL,
  sigma_lbound = NULL,
  sigma_ubound = NULL,
  try = 1000,
  ncores = NULL
)
}
\arguments{
\item{data}{Data frame.
A data frame object of data for potentially
multiple subjects that contain
a column of subject ID numbers
(i.e., an ID variable), and
at least one column of observed values.}

\item{observed}{Character vector.
A vector of character strings
of the names of the observed variables in the data.}

\item{id}{Character string.
A character string of the name of the ID variable in the data.}

\item{time}{Character string.
A character string of the name of the TIME variable in the data.}

\item{phi_start}{Optional starting values for \code{phi}.}

\item{phi_lbound}{Optional lower bound for \code{phi}.}

\item{phi_ubound}{Optional upper bound for \code{phi}.}

\item{sigma_start}{Optional starting values for \code{sigma}.}

\item{sigma_lbound}{Optional lower bound for \code{sigma}.}

\item{sigma_ubound}{Optional upper bound for \code{sigma}.}

\item{try}{Positive integer.
Number of extra tries for \code{\link[OpenMx:mxTryHard]{OpenMx::mxTryHard()}}.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
Fit First Order Continuous-Time Vector Autoregressive Model by ID
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{FitDTVAR}()},
\code{\link{Meta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{fit}
\keyword{metaVAR}
