% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-fit-dt-var.R
\name{FitDTVAR}
\alias{FitDTVAR}
\title{Fit First Order Discrete-Time Vector Autoregressive Model by ID}
\usage{
FitDTVAR(
  data,
  observed,
  id,
  beta_start = NULL,
  beta_lbound = NULL,
  beta_ubound = NULL,
  psi_start = NULL,
  psi_lbound = NULL,
  psi_ubound = NULL,
  try = 1000,
  ncores = NULL
)
}
\arguments{
\item{data}{Data frame.
A data frame object of data for potentially
multiple subjects that contain
a column of subject ID numbers
(i.e., an ID variable), and
at least one column of observed values.}

\item{observed}{Character vector.
A vector of character strings
of the names of the observed variables in the data.}

\item{id}{Character string.
A character string of the name of the ID variable in the data.}

\item{beta_start}{Optional starting values for \code{beta}.}

\item{beta_lbound}{Optional lower bound for \code{beta}.}

\item{beta_ubound}{Optional upper bound for \code{beta}.}

\item{psi_start}{Optional starting values for \code{psi}.}

\item{psi_lbound}{Optional lower bound for \code{psi}.}

\item{psi_ubound}{Optional upper bound for \code{psi}.}

\item{try}{Positive integer.
Number of extra tries for \code{\link[OpenMx:mxTryHard]{OpenMx::mxTryHard()}}.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
Fit First Order Discrete-Time Vector Autoregressive Model by ID
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{FitCTVAR}()},
\code{\link{Meta}()},
\code{\link{Meta.default}()},
\code{\link{Meta.metavarctvar}()},
\code{\link{Meta.metavardtvar}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{fit}
\keyword{metaVAR}
