% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\title{fst}
\usage{
fst(p)
}
\arguments{
\item{p}{allele frequencies (data.frame)}
}
\value{
F-statistics value
}
\description{
Calculate F-statistics from MVS allele frequencies
}
\details{
The F-statistics is computed as follow, \eqn{F=\frac{var(p)}{mean(p)(1-mean(p))}}, where \code{p} is a vector of allele frequencies
}
\references{
Sewall Wright. Genetical Structure of Populations. Nature, 166:247–249, 1950
}
