% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMWIS.R
\name{getMWIS}
\alias{getMWIS}
\title{getMWIS}
\usage{
getMWIS(MVC)
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}
}
\value{
a list of object of class \code{mvc} candidate for \code{mvs}.
}
\description{
getMWIS identifies all the maximum weighted independant sets from a list of \code{mvc}.
}
\examples{
data("MS5")
e = c(3,5)
p = c(5, 10)
MVC = tryParam(e, p , MS5$cov)
MWIS = getMWIS (MVC)
}
\references{
S. Sakai, M. Togasaki, K. Yamazaki. A note on greedy algorithms
for the maximum weighted independent set problem. \emph{Discrete applied Mathematics
}, 126:313-322, 2003.
}
