\name{LK}
\alias{LK}
\title{LK}
\description{Compute the LK statistics}
\usage{
LK(p)
}
\arguments{
  \item{p}{a matrix of allele frequencies}
}
\value{
  \item{LK_val }{LK statistics}
  \item{npop }{number of populations used}
  \item{varID }{variants ID}
}
\references{Lewontin RC, Krakauer J (1973) Distribution of gene frequency as a test of the theory of the selective neutrality of polymorphisms. Genetics 74, 175-195}


