% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-meta.R
\name{Meta}
\alias{Meta}
\title{Fit Multivariate Meta-Analysis}
\usage{
Meta(object, mu_start = NULL, sigma_l_start = NULL, try = 1000, ncores = NULL)
}
\arguments{
\item{object}{Object of class \code{metavardtvar},
that is, the output of \code{\link[=FitDTVAR]{FitDTVAR()}}.}

\item{mu_start}{Numeric matrix.
Matrix of starting values of \code{mu}.}

\item{sigma_l_start}{Numeric matrix.
Matrix of starting values of \code{t(chol(sigma))}.}

\item{try}{Positive integer.
Number of extra tries for \code{\link[OpenMx:mxTryHard]{OpenMx::mxTryHard()}}.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\value{
Returns an object
of class \code{metavarmeta} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{fun}{Function used ("Meta").}
\item{output}{Fitted model.}
\item{transform}{Transformed estimates.}
}
}
\description{
This function estimates the mean and covariance matrix
of a vector of coefficients
using the estimated coefficients and sampling variance-covariance matrix
from each individual.
}
\details{
For \eqn{i = \left\{1, \cdots, n \right\}},
the objective function used to estimate the mean \eqn{\boldsymbol{\mu}}
and covariance matrix \eqn{\boldsymbol{\Sigma}}
of the random coefficients \eqn{\mathbf{y}_{i}} is given by
\deqn{
    \ell
    \left(
    \boldsymbol{\mu} ,
    \boldsymbol{\Sigma} \mid \mathbf{y}_{i},
    \mathbb{V} \left( \mathbf{y}_{i} \right)
    \right)
    =
    - \frac{1}{2}
    \left[
    q \log \left( 2 \pi \right)
    +
    \log
    \left(
    \left|
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right|
    \right)
    +
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)^{\prime}
    \left(
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right)^{-1}
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)
    \right]
  }
where
\eqn{q} is the number of unique elements
in \eqn{\boldsymbol{\mu}} and \eqn{\boldsymbol{\Sigma}}, and
\eqn{\mathbb{V} \left( \mathbf{y}_{i} \right)}
is the sampling variance-covariance matrix of \eqn{\mathbf{y}_{i}}.
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{FitDTVAR}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{meta}
\keyword{metaVAR}
