% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-meta-var-mx.R
\name{MetaVARMx}
\alias{MetaVARMx}
\title{Fit Multivariate Meta-Analysis}
\usage{
MetaVARMx(
  object,
  mu_start = NULL,
  mu_lbound = NULL,
  mu_ubound = NULL,
  sigma_l_start = NULL,
  sigma_l_lbound = NULL,
  sigma_l_ubound = NULL,
  diag = FALSE,
  intercept = FALSE,
  noise = FALSE,
  error = FALSE,
  try = 1000,
  ncores = NULL
)
}
\arguments{
\item{object}{Output of the \code{\link[fitDTVARMx:FitDTVARIDMx]{fitDTVARMx::FitDTVARIDMx()}}
or \code{\link[fitCTVARMx:FitCTVARIDMx]{fitCTVARMx::FitCTVARIDMx()}} functions.}

\item{mu_start}{Numeric vector.
Optional vector of starting values for \code{mu}.}

\item{mu_lbound}{Numeric vector.
Optional vector of lower bound values for \code{mu}.}

\item{mu_ubound}{Numeric vector.
Optional vector of upper bound values for \code{mu}.}

\item{sigma_l_start}{Numeric matrix.
Optional matrix of starting values for \code{t(chol(sigma))}.}

\item{sigma_l_lbound}{Numeric matrix.
Optional matrix of lower bound values for \code{t(chol(sigma))}.}

\item{sigma_l_ubound}{Numeric matrix.
Optional matrix of upper bound values for \code{t(chol(sigma))}.}

\item{diag}{Logical.
If \code{diag = TRUE},
\code{sigma} is a diagonal matrix.
If \code{diag = FALSE},
\code{sigma} is a symmetric matrix.}

\item{intercept}{Logical.
If \code{intercept = TRUE},
include estimates of the process intercept vector, if available.
If \code{intercept = FALSE},
exclude estimates of the process intercept vector.}

\item{noise}{Logical.
If \code{noise = TRUE},
include estimates of the process noise matrix, if available.
If \code{noise = FALSE},
exclude estimates of the process noise matrix.}

\item{error}{Logical.
If \code{error = TRUE},
include estimates of the measurement error matrix, if available.
If \code{error = FALSE},
exclude estimates of the measurement error matrix.}

\item{try}{Positive integer.
Number of extra optimization tries.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
This function estimates the mean and covariance matrix
of a vector of coefficients
using the estimated coefficients and sampling variance-covariance matrix
from each individual.
}
\details{
For \eqn{i = \left\{1, \cdots, n \right\}},
the objective function used to estimate the mean \eqn{\boldsymbol{\mu}}
and covariance matrix \eqn{\boldsymbol{\Sigma}}
of the random coefficients \eqn{\mathbf{y}_{i}} is given by
\deqn{
    \ell
    \left(
    \boldsymbol{\mu} ,
    \boldsymbol{\Sigma} \mid \mathbf{y}_{i},
    \mathbb{V} \left( \mathbf{y}_{i} \right)
    \right)
    =
    - \frac{1}{2}
    \left[
    q \log \left( 2 \pi \right)
    +
    \log
    \left(
    \left|
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right|
    \right)
    +
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)^{\prime}
    \left(
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right)^{-1}
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)
    \right]
  }
where
\eqn{q} is the number of unique elements
in \eqn{\boldsymbol{\mu}} and \eqn{\boldsymbol{\Sigma}}, and
\eqn{\mathbb{V} \left( \mathbf{y}_{i} \right)}
is the sampling variance-covariance matrix of \eqn{\mathbf{y}_{i}}.
}
\examples{
\dontrun{
# Generate data using the simStateSpace package------------------------------
beta_mu <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
beta_sigma <- diag(3 * 3)
beta <- simStateSpace::SimBetaN(
  n = 5,
  beta = beta_mu,
  vcov_beta_vec_l = t(chol(beta_sigma))
)
sim <- simStateSpace::SimSSMVARIVary(
  n = 5,
  time = 100,
  mu0 = list(rep(x = 0, times = 3)),
  sigma0_l = list(t(chol(diag(3)))),
  alpha = list(rep(x = 0, times = 3)),
  beta = beta,
  psi_l = list(t(chol(diag(3))))
)
data <- as.data.frame(sim)

# Fit the model--------------------------------------------------------------
library(fitDTVARMx)
fit <- FitDTVARIDMx(
  data = data,
  observed = c("y1", "y2", "y3"),
  id = "id"
)
# Multivariate meta-analysis-------------------------------------------------
library(metaVAR)
meta <- MetaVARMx(fit)
print(meta)
summary(meta)
coef(meta)
vcov(meta)
}

}
\references{
Neale, M. C., Hunter, M. D., Pritikin, J. N.,
Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R.,
Bates, T. C., Maes, H. H., & Boker, S. M. (2015).
OpenMx 2.0: Extended structural equation and statistical modeling.
\emph{Psychometrika},
\emph{81}(2), 535–549.
\doi{10.1007/s11336-014-9435-8}
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{Meta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{meta}
\keyword{metaVAR}
