% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc_class.R
\docType{class}
\name{mvc-class}
\alias{mvc-class}
\title{Class mvc}
\description{
Class \code{mvc} defines a metavariants cluster.
}
\section{Slots}{

\describe{
   \item{name}{cluster name.}
   \item{eps}{epsilon value used by dbscan.}
   \item{pts}{minimum points value used by dbscan.}
   \item{var}{metavariants IDs.}
   \item{pop}{population names.}
   \item{cov}{depth of coverage matrix of biallelic loci.}
   \item{fit}{\code{fitdist} objects for fitting between observed and theoritical depth of coverage in each population.}
   \item{comp}{component id of the mvc.}
   \item{deg}{mvc degree in mvc graph.}
   \item{weight}{mvc weight.}
   \item{score}{mvc score for maximum weighted independant set}
   \item{mwis}{logical, if true the mvc is a maximum weighted independant set}
 }
}

\author{
momosapiens
}
