% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvCluster.R
\docType{class}
\name{mvCluster-class}
\alias{mvCluster-class}
\title{mvCluster object}
\description{
metavariants cluster
}
\section{Slots}{

\describe{
   \item{name}{cluster name.}
   \item{eps}{epsilon value used by dbscan.}
   \item{pts}{minimum points value used by dbscan.}
   \item{nvar}{number of metavariants}
   \item{pop}{population names where the cluster occurs}
   \item{cov}{depth of coverage matrix of biallelic loci}
   \item{loglik}{log-likelihood of the fitting between observed and theoritical depth of coverage}
 }
}

