% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvs_class.R
\docType{class}
\name{mvs-class}
\alias{mvs-class}
\title{Class mvs}
\description{
Class \code{mvs} defines a metavariants species.
}
\section{Slots}{

\describe{
   \item{id}{mvs id.}
   \item{freq}{allele frequencies}
   \item{gFst}{global Fst}
   \item{gLK}{global LK}
   \item{pwFst}{pairwise-fst}

 }
}

\author{
momosapiens
}
