% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvs.R
\name{mvs}
\alias{mvs}
\title{mvs}
\usage{
mvs(id = "", freq = data.frame(), gFst = 0, gLK = 0,
  pwFst = data.frame(), name = "", eps = 0, pts = 0, var = c(),
  pop = c(), cov = data.frame(), fit = list(), comp = 0, deg = 0,
  weight = 0, score = 0, mwis = FALSE)
}
\arguments{
\item{id}{mvs id}

\item{freq}{allele frequency}

\item{gFst}{global Fst}

\item{gLK}{global LK}

\item{pwFst}{pairwise-Fst}

\item{name}{cluster name.}

\item{eps}{epsilon value used by dbscan.}

\item{pts}{minimum points value used by dbscan.}

\item{var}{metavariants IDs.}

\item{pop}{population names.}

\item{cov}{depth of coverage matrix of biallelic loci.}

\item{fit}{\code{fitdist} object for fitting between observed and theoritical depth of coverage.}

\item{comp}{component id of the mvc.}

\item{deg}{mvc degree in the mvc graph}

\item{weight}{mvc weight}

\item{score}{mvc score for maximum weighted independant set}

\item{mwis}{logical, set to true if the mvc is a maximum weighted independant set}
}
\description{
mvs
}
