% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwis.R
\name{mwis}
\alias{mwis}
\title{mwis}
\usage{
mwis(nbComp, MVC)
}
\arguments{
\item{nbComp}{number of connected components}

\item{MVC}{a list of object of class \code{mvc}.}
}
\value{
a list of object of class \code{mvc}.
}
\description{
Calculate the mvc weight and score in a list of object of class \code{mvc}.
Identify some the maximum weighted independant sets and change the slot mwis
}
