% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMvs.R
\docType{methods}
\name{plotMvs}
\alias{plotMvs}
\alias{plotMvs,mvs-method}
\title{plotMvs}
\usage{
plotMvs(mvs, type)

\S4method{plotMvs}{mvs}(mvs, type)
}
\arguments{
\item{mvs}{an object of class \code{mvs}.}

\item{type}{type of plot to produce}
}
\value{
none
}
\description{
plotMvs allows various plotting for \code{mvs} like depth of coverage and allele frequency distribution, pairwise F-statistics, LK distribution
}
\details{
The type of plot can be \code{freq} for allele frequencies distribution,
\code{cov} for depth of coverage distribution.
\code{heatFst} for a heatmap of the pairwise F-statistics
\code{LK} for the observed and theoritical LK statistics
}
