\name{plotPW}
\alias{plotPW}
\title{plotPW}
\description{Plot the pairwise Fst matrix as a tree}
\usage{plotPW(mfst)}
\arguments{
  \item{mfst}{a pairwise Fst matrix}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\seealso{
plotAF, plotLK
}
\examples{

}
