% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarctvar.R
\name{print.metavarctvar}
\alias{print.metavarctvar}
\title{Print Method for Object of Class \code{metavarctvar}}
\usage{
\method{print}{metavarctvar}(x, means = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{metavarctvar}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{metavarctvar}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
