% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavardtvar.R
\name{print.metavardtvar}
\alias{print.metavardtvar}
\title{Print Method for Object of Class \code{metavardtvar}}
\usage{
\method{print}{metavardtvar}(x, means = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{metavardtvar}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{metavardtvar}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
