% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarmeta.R
\name{print.metavarmeta}
\alias{print.metavarmeta}
\title{Print Method for Object of Class \code{metavarmeta}}
\usage{
\method{print}{metavarmeta}(x, alpha = 0.05, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{metavarmeta}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{metavarmeta}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
