% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMvs.R
\name{readMvs}
\alias{readMvs}
\title{readMvs}
\usage{
readMvs(mvsDir)
}
\arguments{
\item{mvsDir}{output directory}
}
\value{
an object of class \code{mvs}
}
\description{
Read a \code{mvs} from a directory dreated by \code{writeMvs}
}
