% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMvsList.R
\name{readMvsList}
\alias{readMvsList}
\title{readMvsList}
\usage{
readMvsList(dir = "")
}
\arguments{
\item{dir}{a directory.}
}
\value{
a list of \code{mvs}
}
\description{
read \code{mvs} from a list of directory
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]

# MVC and MVS creation
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)

# store and read MVC
writeMvcList(MVC, "MVC")
MVC2 = readMvcList("MVC")

# store and read MVS
writeMvsList(MVS, "MVS")
MVS2 = readMvsList("MVS")
unlink("MVC")
unlink("MVS")

}
}
