% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMvc.R
\docType{methods}
\name{setMvc}
\alias{setMvc}
\alias{setMvc,mvc-method}
\title{setMvc}
\usage{
setMvc(
  x,
  name = x@name,
  eps = x@eps,
  pts = x@pts,
  var = x@var,
  pop = x@pop,
  cov = x@cov,
  fit = x@fit,
  comp = x@comp,
  deg = x@deg,
  weight = x@weight,
  score = x@score,
  mwis = x@mwis
)

\S4method{setMvc}{mvc}(
  x,
  name = x@name,
  eps = x@eps,
  pts = x@pts,
  var = x@var,
  pop = x@pop,
  cov = x@cov,
  fit = x@fit,
  comp = x@comp,
  deg = x@deg,
  weight = x@weight,
  score = x@score,
  mwis = x@mwis
)
}
\arguments{
\item{x}{an object of class \code{mvc}.}

\item{name}{cluster name.}

\item{eps}{epsilon value used by dbscan.}

\item{pts}{minimum points value used by dbscan.}

\item{var}{metavariants IDs.}

\item{pop}{population names.}

\item{cov}{depth of coverage matrix of biallelic loci.}

\item{fit}{\code{fitdist} object for fitting between observed and theoritical depth of coverage.}

\item{comp}{component id of the mvc.}

\item{deg}{mvc degree in the mvc graph}

\item{weight}{mvc weight}

\item{score}{mvc score for maximum weighted independant set}

\item{mwis}{logical, if true the mvc is a maximum weighted independant set}
}
\value{
an object of class \code{mvc}.
}
\description{
set \code{mvc} attributes
}
