% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMvs.R
\docType{methods}
\name{setMvs}
\alias{setMvs}
\alias{setMvs,mvs-method}
\title{setMvs}
\usage{
setMvs(
  x,
  id = x@id,
  freq = x@freq,
  gFst = x@gFst,
  gLK = x@gLK,
  pwFst = x@pwFst,
  name = x@name,
  eps = x@eps,
  pts = x@pts,
  var = x@var,
  pop = x@pop,
  cov = x@cov,
  fit = x@fit,
  comp = x@comp,
  deg = x@deg,
  weight = x@weight,
  score = x@score,
  mwis = x@mwis
)

\S4method{setMvs}{mvs}(
  x,
  id = x@id,
  freq = x@freq,
  gFst = x@gFst,
  gLK = x@gLK,
  pwFst = x@pwFst,
  name = x@name,
  eps = x@eps,
  pts = x@pts,
  var = x@var,
  pop = x@pop,
  cov = x@cov,
  fit = x@fit,
  comp = x@comp,
  deg = x@deg,
  weight = x@weight,
  score = x@score,
  mwis = x@mwis
)
}
\arguments{
\item{x}{an object of class \code{mvc}.}

\item{id}{mvs name.}

\item{freq}{allele frequencies}

\item{gFst}{global Fst}

\item{gLK}{global LK}

\item{pwFst}{pairwise-Fst}

\item{name}{cluster name.}

\item{eps}{epsilon value used by dbscan.}

\item{pts}{minimum points value used by dbscan.}

\item{var}{metavariants IDs.}

\item{pop}{population names.}

\item{cov}{depth of coverage matrix of biallelic loci.}

\item{fit}{\code{fitdist} object for fitting between observed and theoritical depth of coverage.}

\item{comp}{component id of the mvc.}

\item{deg}{mvc degree in the mvc graph}

\item{weight}{mvc weight}

\item{score}{mvc score for maximum weighted independant set}

\item{mwis}{logical, if true the mvc is a maximum weighted independant set}

\item{mvs}{an object of class \code{mvs}.}
}
\value{
an object of class \code{mvs}.
}
\description{
set \code{mvs} attributes
}
