% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarctvar.R
\name{summary.metavarctvar}
\alias{summary.metavarctvar}
\title{Summary Method for Object of Class \code{metavarctvar}}
\usage{
\method{summary}{metavarctvar}(object, means = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{metavarctvar}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{metavarctvar}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
