% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavardtvar.R
\name{summary.metavardtvar}
\alias{summary.metavardtvar}
\title{Summary Method for Object of Class \code{metavardtvar}}
\usage{
\method{summary}{metavardtvar}(object, means = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{metavardtvar}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{metavardtvar}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
