% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryParam.R
\name{tryParam}
\alias{tryParam}
\title{tryParam}
\usage{
tryParam(eps, pts, cov, mvcMinVar = 1000)
}
\arguments{
\item{eps}{epsilon value(s).}

\item{pts}{minimum points value(s).}

\item{cov}{depth of coverage of biallelic loci in \code{data.frame}.}

\item{mvcMinVar}{minimum of metavariants in \code{mcv}, set to 1,000 by default.}
}
\value{
a list of objects of class \code{mcv}.
}
\description{
tryParam tests several couples of epsilon and minimum points values for density-based clustering of metavariants.
}
\examples{
data("MS5")
e = c(3,5)
p = c(5, 10)
MVC = tryParam(e, p , MS5$cov) # tests (3,5), (3,10), (5,5), (5, 10)


}
