% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFilterbyCov.R
\name{varFilterbyCov}
\alias{varFilterbyCov}
\title{varFilterbyCov}
\usage{
varFilterbyCov(cov, dev = 2, minCov = 8)
}
\arguments{
\item{cov}{depth of coverage}

\item{dev}{number of standard deviation}

\item{minCov}{minimum depth of coverage}
}
\description{
varFilterbyCov
}
