% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods-metavarmeta.R
\name{vcov.metavarmeta}
\alias{vcov.metavarmeta}
\title{Variance-Covariance Matrix Method for an Object of Class
\code{metavarmeta}}
\usage{
\method{vcov}{metavarmeta}(object, ...)
}
\arguments{
\item{object}{an object of class \code{metavarmeta}.}

\item{...}{further arguments.}
}
\value{
Returns the variance-covariance matrix of the estimated parameters.
}
\description{
Variance-Covariance Matrix Method for an Object of Class
\code{metavarmeta}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
