% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMvc.R
\docType{methods}
\name{writeMvc}
\alias{writeMvc}
\alias{writeMvc,mvc-method}
\title{writeMvc}
\usage{
writeMvc(x, prefix = "out")

\S4method{writeMvc}{mvc}(x, prefix = "out")
}
\arguments{
\item{x}{an object of class \code{mvc}.}

\item{prefix}{output directory}
}
\description{
Write a \code{mvc} in an output directory
}
