% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMvcList.R
\name{writeMvcList}
\alias{writeMvcList}
\title{writeMvcList}
\usage{
writeMvcList(MVC, prefix = "out")
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}

\item{prefix}{output directory}
}
\description{
Write a list of \code{mvc} in an output directory
}
