% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMvs.R
\docType{methods}
\name{writeMvs}
\alias{writeMvs}
\alias{writeMvs,mvs-method}
\alias{writeMvs,mvc-method}
\title{write a object Mvs}
\usage{
writeMvs(x, prefix = "out")

\S4method{writeMvs}{mvs}(x, prefix = "out")
}
\arguments{
\item{x}{an object of class \code{mvs}.}

\item{prefix}{output directory}
}
\description{
Write a \code{mvs} in an output directory
}
