% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMvsList.R
\name{writeMvsList}
\alias{writeMvsList}
\title{writeMvsList}
\usage{
writeMvsList(MVS, prefix = "out")
}
\arguments{
\item{MVS}{a list of objects of class \code{mvs}.}

\item{prefix}{output directory}
}
\description{
Write a list of \code{mvs} in an output directory
}
