\name{Fst}
\alias{Fst}
\title{Fst}
\description{Compute the global Fst from a matrix a allele frequencies}
\usage{Fst(p)}

\arguments{
  \item{p}{matrix of allele frequencies with variants in row and populations in column}
}
\details{}
\value{

  \item{fst }{fst values}
  \item{varID }{variants ID}
}
\references{Wright S. (1951) The genetical structure of populations. Ann Eugen 15, 323-354.}
\author{}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}

