% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fst}
\alias{fst}
\title{F-statistics}
\usage{
fst(p)
}
\arguments{
\item{p}{allele frequencies of class \code{data.frame}}
}
\value{
\code{vector} of F-statistics values
}
\description{
Compute F-statistics from MVS allele frequencies
}
\details{
The F-statistics is computed as follow, \eqn{F=\frac{var(p)}{mean(p)(1-mean(p))}}, where \code{p} is a vector of allele frequencies
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]

# Testing dbscan parameters
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)

# F-statistics
FST =fst(MVS[[1]]@freq)


}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)

# Create and MVC and MVS
MVC = tryParam (e, p, MS5$cov)
MWIS = getMWIS (MVC)
MVS = mvc2mvs(MWIS, freq = MS5$freq)

# F-statistics
FST = Fst(MVS[[1]]@freq)
hist(FST)
}
}
\references{
Sewall Wright. Genetical Structure of Populations. Nature, 166:247–249, 1950
}
