% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMWIS.R
\name{getMWIS}
\alias{getMWIS}
\title{getMWIS}
\usage{
getMWIS(MVC, p = FALSE, prefix = "out")
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}

\item{p}{\code{logical}, for graphical output, set to TRUE by default.}

\item{prefix}{name of the graphical output directory, set to "out" by default.}
}
\value{
a list of object of class \code{mvc} candidate for \code{mvs}.
}
\description{
getMWIS identifies all the maximum weighted independant sets from a list of \code{mvc}.
}
\examples{
\dontrun{data("MS5")
e = c(5,6)
p = c(5, 10)
MVC = tryParam(e, p , MS5$cov)
MWIS = getMWIS (MVC)}
}
\references{
S. Sakai, M. Togasaki, K. Yamazaki. A note on greedy algorithms
for the maximum weighted independent set problem. \emph{Discrete applied Mathematics
}, 126:313-322, 2003.
}
