% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMWIS.R
\name{getMWIS}
\alias{getMWIS}
\title{getMWIS}
\usage{
getMWIS(MVC, p = FALSE, prefix = "out")
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}

\item{p}{\code{logical}, for graphical output, set to TRUE by default.}

\item{prefix}{name of the graphical output directory, set to "out" by default.}
}
\value{
a list of object of class \code{mvc} candidate for \code{mvs}.
}
\description{
getMWIS identifies all the maximum weighted independant sets from a list of \code{mvc}.
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]
# Testing dbscan parameters
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)
plotMvs(MVS[[1]], type = "heatFst")
}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# Testing dbscan parameters
MVC = tryParam (e, p, MS5$cov)
MWIS = getMWIS (MVC)
MVS = mvc2mvs(MWIS, freq = MS5$freq)
plotMvs(MVS[[1]], type = "heatFst")
}
}
\references{
S. Sakai, M. Togasaki, K. Yamazaki. A note on greedy algorithms
for the maximum weighted independent set problem. \emph{Discrete applied Mathematics
}, 126:313-322, 2003.
}
