% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc2AdjMat.R
\name{graph2mvcComp}
\alias{graph2mvcComp}
\title{graph2mvcComp}
\usage{
graph2mvcComp(graph, mvcList)
}
\arguments{
\item{graph}{an adjacency matrix.}

\item{mvcList}{a list of object of class \code{mvc}.}
}
\value{
a list of object of class \code{mvc} assigned to graph component.
}
\description{
Extract connected mvs from the mvc graph and attribute the mvc to components
}
