% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LK.R
\name{LK}
\alias{LK}
\title{LK}
\usage{
LK(p)
}
\arguments{
\item{p}{allele frequencies of class \code{data.frame}}
}
\value{
LK value of class \code{data.frame}
}
\description{
Compute the LK statistics from MVS allele frequencies
}
\details{
The LK is computed as follow, \eqn{LK=\frac{(n-1) Fst}{mean(F_{ST})}}, where \code{F_{ST}} is the wright's F-statistics.
}
\references{
RC Lewontin and J Krakauer. Distribution of gene frequency as a test of the theory of the selective neutrality of polymorphisms. Genetics 74(1):175–195. 1973
}
