% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-meta.R
\name{Meta}
\alias{Meta}
\title{Fit Multivariate Meta-Analysis}
\usage{
Meta(
  y,
  v,
  mu_start = NULL,
  mu_lbound = NULL,
  mu_ubound = NULL,
  sigma_l_start = NULL,
  sigma_l_lbound = NULL,
  sigma_l_ubound = NULL,
  diag = FALSE,
  try = 1000,
  ncores = NULL
)
}
\arguments{
\item{y}{A list.
Each element of the list is a numeric vector
of estimated coefficients.}

\item{v}{A list.
Each element of the list
is a sampling variance-covariance matrix of \code{y}.}

\item{mu_start}{Numeric vector.
Optional vector of starting values for \code{mu}.}

\item{mu_lbound}{Numeric vector.
Optional vector of lower bound values for \code{mu}.}

\item{mu_ubound}{Numeric vector.
Optional vector of upper bound values for \code{mu}.}

\item{sigma_l_start}{Numeric matrix.
Optional matrix of starting values for \code{t(chol(sigma))}.}

\item{sigma_l_lbound}{Numeric matrix.
Optional matrix of lower bound values for \code{t(chol(sigma))}.}

\item{sigma_l_ubound}{Numeric matrix.
Optional matrix of upper bound values for \code{t(chol(sigma))}.}

\item{diag}{Logical.
If \code{diag = TRUE},
\code{sigma} is a diagonal matrix.
If \code{diag = FALSE},
\code{sigma} is a symmetric matrix.}

\item{try}{Positive integer.
Number of extra optimization tries.}

\item{ncores}{Positive integer.
Number of cores to use.}
}
\description{
This function estimates the mean and covariance matrix
of a vector of coefficients
using the estimated coefficients and sampling variance-covariance matrix
from each individual.
}
\details{
For \eqn{i = \left\{1, \cdots, n \right\}},
the objective function used to estimate the mean \eqn{\boldsymbol{\mu}}
and covariance matrix \eqn{\boldsymbol{\Sigma}}
of the random coefficients \eqn{\mathbf{y}_{i}} is given by
\deqn{
    \ell
    \left(
    \boldsymbol{\mu} ,
    \boldsymbol{\Sigma} \mid \mathbf{y}_{i},
    \mathbb{V} \left( \mathbf{y}_{i} \right)
    \right)
    =
    - \frac{1}{2}
    \left[
    q \log \left( 2 \pi \right)
    +
    \log
    \left(
    \left|
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right|
    \right)
    +
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)^{\prime}
    \left(
      \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma}
    \right)^{-1}
    \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)
    \right]
  }
where
\eqn{q} is the number of unique elements
in \eqn{\boldsymbol{\mu}} and \eqn{\boldsymbol{\Sigma}}, and
\eqn{\mathbb{V} \left( \mathbf{y}_{i} \right)}
is the sampling variance-covariance matrix of \eqn{\mathbf{y}_{i}}.
}
\references{
Neale, M. C., Hunter, M. D., Pritikin, J. N.,
Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R.,
Bates, T. C., Maes, H. H., & Boker, S. M. (2015).
OpenMx 2.0: Extended structural equation and statistical modeling.
\emph{Psychometrika},
\emph{81}(2), 535–549.
\doi{10.1007/s11336-014-9435-8}
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{MetaVARMx}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{meta}
\keyword{metaVAR}
