% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-package.R
\docType{package}
\name{metaVAR-package}
\alias{metaVAR}
\alias{metaVAR-package}
\title{metaVAR: Multivariate Meta-Analysis of Vector Autoregressive Model Estimates}
\description{
Fits fixed-, random-, or mixed-effects multivariate meta-analysis models using vector autoregressive model estimates from each individual.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/metaVAR}
  \item \url{https://jeksterslab.github.io/metaVAR/}
  \item Report bugs at \url{https://github.com/jeksterslab/metaVAR/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
