% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{metaVaR}
\alias{metaVaR}
\alias{MS5}
\alias{metaVaR-package}
\title{metavariants data}
\format{a list of object of class \code{data.frame}
\describe{
 \item{pop}{population names}
 \item{cov}{biallelic loci depth of coverage}
 \item{freq}{metavariant frequencies}
}}
\source{
Genoscope and Tara Oceans consortium
}
\usage{
data(MS5)
}
\description{
Data from discoSNP++ ran on five samples of metagenomic data genrated by the Tara Oceans expedition in the Mediterranean Sea
}
\keyword{datasets}
