% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-meta-var-mx.R
\name{MetaVARMx}
\alias{MetaVARMx}
\title{Fit Multivariate Meta-Analysis}
\usage{
MetaVARMx(
  object,
  x = NULL,
  beta0_values = NULL,
  beta0_free = NULL,
  beta0_lbound = NULL,
  beta0_ubound = NULL,
  beta1_values = NULL,
  beta1_free = NULL,
  beta1_lbound = NULL,
  beta1_ubound = NULL,
  tau_values = NULL,
  tau_free = NULL,
  tau_lbound = NULL,
  tau_ubound = NULL,
  random = TRUE,
  diag = FALSE,
  intercept = FALSE,
  noise = FALSE,
  error = FALSE,
  try = 1000,
  ncores = NULL,
  ...
)
}
\arguments{
\item{object}{Output of the \code{\link[fitDTVARMx:FitDTVARIDMx]{fitDTVARMx::FitDTVARIDMx()}}
or \code{\link[fitCTVARMx:FitCTVARIDMx]{fitCTVARMx::FitCTVARIDMx()}} functions.}

\item{x}{An optional list.
Each element of the list is a numeric vector
of covariates for the mixed-effects model.}

\item{beta0_values}{Numeric vector.
Optional vector of starting values for \code{beta0}.}

\item{beta0_free}{Logical vector.
Optional vector of free (\code{TRUE}) parameters for \code{beta0}.}

\item{beta0_lbound}{Numeric vector.
Optional vector of lower bound values for \code{beta0}.}

\item{beta0_ubound}{Numeric vector.
Optional vector of upper bound values for \code{beta0}.}

\item{beta1_values}{Numeric matrix.
Optional matrix of starting values for \code{beta1}.}

\item{beta1_free}{Logical matrix.
Optional matrix of free (\code{TRUE}) parameters for \code{beta1}.}

\item{beta1_lbound}{Numeric matrix.
Optional matrix of lower bound values for \code{beta1}.}

\item{beta1_ubound}{Numeric matrix.
Optional matrix of upper bound values for \code{beta1}.}

\item{tau_values}{Numeric matrix.
Optional matrix of starting values for \code{t(chol(tau_sqr))}.}

\item{tau_free}{Numeric matrix.
Optional matrix of free (\code{TRUE}) parameters for \code{t(chol(tau_sqr))}.}

\item{tau_lbound}{Numeric matrix.
Optional matrix of lower bound values for \code{t(chol(tau_sqr))}.}

\item{tau_ubound}{Numeric matrix.
Optional matrix of upper bound values for \code{t(chol(tau_sqr))}.}

\item{random}{Logical.
If \code{random = TRUE},
estimates random effects.
If \code{random = FALSE},
\code{tau_sqr} is a null matrix.}

\item{diag}{Logical.
If \code{diag = TRUE},
\code{tau_sqr} is a diagonal matrix.
If \code{diag = FALSE},
\code{tau_sqr} is a symmetric matrix.}

\item{intercept}{Logical.
If \code{intercept = TRUE},
include estimates of the process intercept vector, if available.
If \code{intercept = FALSE},
exclude estimates of the process intercept vector.}

\item{noise}{Logical.
If \code{noise = TRUE},
include estimates of the process noise matrix, if available.
If \code{noise = FALSE},
exclude estimates of the process noise matrix.}

\item{error}{Logical.
If \code{error = TRUE},
include estimates of the measurement error matrix, if available.
If \code{error = FALSE},
exclude estimates of the measurement error matrix.}

\item{try}{Positive integer.
Number of extra optimization tries.}

\item{ncores}{Positive integer.
Number of cores to use.}

\item{...}{Additional optional arguments to pass to \code{mxTryHardctsem}.}
}
\description{
This function estimates
fixed-, random-, or mixed-effects meta-analysis parameters
using the estimated coefficients and sampling variance-covariance matrix
from each individual fitted using the \code{\link[fitDTVARMx:FitDTVARIDMx]{fitDTVARMx::FitDTVARIDMx()}}
or \code{\link[fitCTVARMx:FitCTVARIDMx]{fitCTVARMx::FitCTVARIDMx()}} functions.
}
\examples{
\dontrun{
# Generate data using the simStateSpace package------------------------------
beta_mu <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
beta_sigma <- diag(3 * 3)
beta <- simStateSpace::SimBetaN(
  n = 5,
  beta = beta_mu,
  vcov_beta_vec_l = t(chol(beta_sigma))
)
sim <- simStateSpace::SimSSMVARIVary(
  n = 5,
  time = 100,
  mu0 = list(rep(x = 0, times = 3)),
  sigma0_l = list(t(chol(diag(3)))),
  alpha = list(rep(x = 0, times = 3)),
  beta = beta,
  psi_l = list(t(chol(diag(3))))
)
data <- as.data.frame(sim)

# Fit the model--------------------------------------------------------------
library(fitDTVARMx)
fit <- FitDTVARIDMx(
  data = data,
  observed = c("y1", "y2", "y3"),
  id = "id"
)

# Multivariate meta-analysis-------------------------------------------------
library(metaVAR)
meta <- MetaVARMx(fit)
print(meta)
summary(meta)
coef(meta)
vcov(meta)
}

}
\references{
Cheung, M. W.-L. (2015).
\emph{Meta-analysis: A structural equation modeling approach}.
Chichester, West Sussex: John Wiley & Sons, Inc.

Neale, M. C., Hunter, M. D., Pritikin, J. N.,
Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R.,
Bates, T. C., Maes, H. H., & Boker, S. M. (2015).
OpenMx 2.0: Extended structural equation and statistical modeling.
\emph{Psychometrika},
\emph{81}(2), 535–549.
\doi{10.1007/s11336-014-9435-8}
}
\seealso{
Other Meta-Analysis of VAR Functions: 
\code{\link{Meta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Meta-Analysis of VAR Functions}
\keyword{meta}
\keyword{metaVAR}
