% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc2AdjMat.R
\name{mvc2AdjMat}
\alias{mvc2AdjMat}
\title{mvc2AdjMat}
\usage{
mvc2AdjMat(MVC)
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}
}
\value{
a mvc adjacency matrix.
}
\description{
Create an adjacency matrix mvc based on shared metavariants between \code{mvc}
}
