% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc2AdjMat.R
\name{mvc2graph}
\alias{mvc2graph}
\title{mvc2graph}
\usage{
mvc2graph(mvcList)
}
\arguments{
\item{mvcList}{a list of object of class \code{mvc}.}
}
\value{
a mvc graph of class \code{igraph}.
}
\description{
Create an undirected graph of a list of \code{mvc}
}
