% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvc2mvs.R
\name{mvc2mvs}
\alias{mvc2mvs}
\title{mvc2mvs}
\usage{
mvc2mvs(MVC, minPop = 3, minCov = 8, freq, minVarCov = 8, sd = 2,
  minVar = 100)
}
\arguments{
\item{MVC}{a list of object of class \code{mvc}.}

\item{minPop}{minimum number of population with \code{minCov}, set to 3 by default.}

\item{minCov}{minimum median depth of coverage, set to 8 by default.}

\item{freq}{allele frequencies}

\item{minVarCov}{minimum depth of coverage of loci, set to 8 by default.}

\item{sd}{standard deviation of depth of coverage to select variant, set to 2 by default.}

\item{minVar}{minimum number of variants in a \code{mvs}, set to 100 by default}
}
\value{
a list of objects of class \code{mvs}.
}
\description{
mvc2mvs selects the \code{mvc} belonging to the maximum weighted independent
set and passing quality filters to generates the \code{mvs}. The
}
