% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMvs.R
\docType{methods}
\name{plotMvs}
\alias{plotMvs}
\alias{plotMvs,mvs-method}
\title{plotMvs}
\usage{
plotMvs(mvs, type)

\S4method{plotMvs}{mvs}(mvs, type)
}
\arguments{
\item{mvs}{an object of class \code{mvs}.}

\item{type}{type of plot to produce}
}
\description{
plotMvs allows various plotting for \code{mvs}
}
\details{
type can be "freq" for allele frequencies distribution,
"cov", for depth of coverage distribution.
"heatFst", for a heatmap of the pairwise-Fst
"graphFst", for the "gFst"
}
