% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMvs.R
\docType{methods}
\name{plotMvs}
\alias{plotMvs}
\alias{plotMvs,mvs-method}
\title{plotMvs}
\usage{
plotMvs(x, t)

\S4method{plotMvs}{mvs}(x, t)
}
\arguments{
\item{x}{an object of class \code{mvs}.}

\item{t}{type of plot to produce, can "freq", "cov", "heatFst", "graphFst", "gFst"}
}
\description{
plotMvs allows various plotting of a \code{mvs}
}
