% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFst.R
\name{pwFst}
\alias{pwFst}
\title{pairwise F-statistics}
\usage{
pwFst(p)
}
\arguments{
\item{p}{allele frequencies (data.frame) with row and column names}
}
\value{
\code{data.frame} of symetric pairwise F-statistics
}
\description{
Calculate pairwise F-statistics between each population of a MVS
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]

# MVC and MVS creation
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)

# pairwise-FST
pwFst(MVS[[1]]@freq)

}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)

# MVC and MVS creation
MVC = tryParam (e, p, MS5$cov)
MWIS = getMWIS (MVC)
MVS = mvc2mvs(MWIS, freq = MS5$freq)

# Pairwie-FST
pwFst(MVS[[1]]@freq)
plotMvs(MVS[[1]], type = "heatFst")

}
}
\references{
Sewal Wright. Genetical Structure of Populations. Nature, 166:247–249, 1950
}
