% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwFst.R
\name{pwFst}
\alias{pwFst}
\title{pwFst}
\usage{
pwFst(p)
}
\arguments{
\item{p}{allele frequencies}
}
\value{
pairwise Fst
}
\description{
Calculate pairwise F-statistics between each population
}
\references{
Sewal Wright. Genetical Structure of Populations. Nature, 166:247–249, 1950
}
