% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMvc.R
\name{readMvc}
\alias{readMvc}
\title{readMvc}
\usage{
readMvc(mvcDir)
}
\arguments{
\item{mvcDir}{output directory}
}
\value{
an object of class \code{mvc}
}
\description{
Read a \code{mvc} from a directory dreated by \code{writeMvc}
}
