% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMvcList.R
\name{readMvcList}
\alias{readMvcList}
\title{readMvcList}
\usage{
readMvcList(dir = "")
}
\arguments{
\item{dir}{a directory.}
}
\value{
a list of \code{mvc}
}
\description{
read \code{mvc} from a list of directory
}
