% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMvsList.R
\name{readMvsList}
\alias{readMvsList}
\title{readMvsList}
\usage{
readMvsList(dir = "")
}
\arguments{
\item{dir}{a directory.}
}
\value{
a list of \code{mvs}
}
\description{
read \code{mvs} from a list of directory
}
