% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVAR-methods.R
\name{summary.metavarmeta}
\alias{summary.metavarmeta}
\title{Summary Method for Object of Class \code{metavarmeta}}
\usage{
\method{summary}{metavarmeta}(object, alpha = 0.05, digits = 4, ...)
}
\arguments{
\item{object}{an object of class \code{metavarmeta}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.}

\item{digits}{Integer indicating the number of decimal places to display.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{metavarmeta}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
