% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryParam.R
\name{tryParam}
\alias{tryParam}
\title{tryParam}
\usage{
tryParam(eps, pts, cov, mvcMinVar = 1000)
}
\arguments{
\item{eps}{epsilon value(s).}

\item{pts}{minimum points value(s).}

\item{cov}{depth of coverage of biallelic loci in \code{data.frame}.}

\item{mvcMinVar}{minimum of metavariants in \code{mcv}, set to 1,000 by default.}
}
\value{
a list of objects of class \code{mcv}.
}
\description{
tryParam tests several couples of epsilon and minimum points values for density-based clustering of metavariants.
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]
# Testing dbscan parameters
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)
plotMvs(MVS[[1]], type = "heatFst")
}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# Testing dbscan parameters
MVC = tryParam (e, p, MS5$cov)
MWIS = getMWIS (MVC)
MVS = mvc2mvs(MWIS, freq = MS5$freq)
plotMvs(MVS[[1]], type = "heatFst")
}
}
