% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryParam.R
\name{tryParam}
\alias{tryParam}
\title{tryParam
Test several couples of epsilon and minimum points for density-based clustering of metavariants}
\usage{
tryParam(eps, pts, cov, mvcMinVar = 1000)
}
\arguments{
\item{eps}{epsilon value(s).}

\item{pts}{minimum points value(s).}

\item{cov}{dataframe of the depth of coverage of biallelic loci.}

\item{mvcMinVar}{minimum of metavariants in \code{mcv}, set to 1,000 by default.}
}
\value{
a list of object of class \code{mcv}
}
\description{
tryParam
Test several couples of epsilon and minimum points for density-based clustering of metavariants
}
\examples{
{
data("MS5")
e = c(3,4)
p = c(10, 20)
#test couples (3,10), (4,10), (3, 20), (4,20)
}

}
