% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFilterbyCov.R
\name{varFilterbyCov}
\alias{varFilterbyCov}
\title{Loci fitering by coberage}
\usage{
varFilterbyCov(cov, dev = 2, minCov = 8)
}
\arguments{
\item{cov}{depth of coverage}

\item{dev}{number of standard deviation}

\item{minCov}{minimum depth of coverage}
}
\value{
depth of coverage
}
\description{
Filtering of variable loci dependending on the expected depth of coverage
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]
# Testing dbscan parameters
MVC = tryParam (e, p, coverage)
Cov = varFilterbyCov (MVC[[1]]@cov)
}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# Testing dbscan parameters
MVC = tryParam (e, p, MS5$cov)
Cov = varFilterbyCov (MVC[[1]]@cov)
}
}
