% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMvcList.R
\name{writeMvcList}
\alias{writeMvcList}
\title{write a Mvc list}
\usage{
writeMvcList(MVC, prefix = "../out")
}
\arguments{
\item{MVC}{a list of objects of class \code{mvc}.}

\item{prefix}{output directory (default is out)}
}
\description{
writeMvcList Writes a list of object of class \code{mvc} in at directory
}
\examples{
\dontshow{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)
# sampling 1000 loci from the Mediterranean variant set
loci = sample(rownames(MS5$cov), 10000)
coverage = MS5$cov[loci,]

# MVC and MVS creation
MVC = tryParam (e, p, coverage)
MWIS = getMWIS (MVC)
frequencies = MS5$freq[loci,]
MVS = mvc2mvs(MWIS, freq = frequencies)

# store MVC
writeMvcList(MVC, "MVC")

# store MVS
writeMvsList(MVS, "MVS")

unlink("MVC")
unlink("MVS")

}
\donttest{
# espilon values to test for dbscan
e = c(3,5)
# minimum points values to test for dbscan
p = c(5,10)

# MVC and MVS creation
MVC = tryParam (e, p, MS5$cov)
MWIS = getMWIS (MVC)
MVS = mvc2mvs(MWIS, freq = MS5$freq)

# Pairwie-FST
pwFst(MVS[[1]]@freq)
plotMvs(MVS[[1]], type = "heatFst")

# store MVC
writeMvcList(MVC, "MVC")

# store MVS
writeMvsList(MVS, "MVS")

}
}
